mv location tracker to combine with LDS wrapper as WorldTrackingCameraController
